Trying out new way of hiding mouse
The old version of this extension hides the mouse when it is idle and shows it when you move it. This works perfectly fine, but it does not alway match up with the visibility of the web player ui. This new algorithm mirrors the visibility of the cursor with the visibiliy of the controls.